Move style source files to src and build to dist
