+++ styles/package.json
@@ -1,22 +1,21 @@
 {
-  "name": "styles",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "build": "ts-node buildThemes.ts",
-    "watch": "nodemon"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/chroma-js": "^2.1.3",
-    "@types/node": "^17.0.23",
-    "case-anything": "^2.1.10",
-    "chroma-js": "^2.4.2",
-    "ts-node": "^10.7.0"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.15" 
-  }
+    "name": "styles",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "build-themes": "ts-node ./src/buildThemes.ts",
+        "build-figma": "ts-node ./src/buildFigmaTokens.ts",
+        "watch": "nodemon"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@types/chroma-js": "^2.1.3",
+        "@types/node": "^17.0.23",
+        "case-anything": "^2.1.10",
+        "chroma-js": "^2.4.2",
+        "ts-node": "^10.7.0",
+        "nodemon": "^2.0.15"
+    }
 }
