+++ styles/src/buildThemes.ts
@@ -3,14 +3,14 @@ import * as path from "path";
 import app from "./styleTree/app";
 import dark from "./themes/dark";
 import light from "./themes/light";
-import decamelizeTree from "./utils/decamelizeTree";
+import snakeCase from "./utils/snakeCase";
 
 const themes = [dark, light];
 for (let theme of themes) {
-    let styleTree = decamelizeTree(app(theme));
+    let styleTree = snakeCase(app(theme));
     let styleTreeJSON = JSON.stringify(styleTree, null, 2);
     let outPath = path.resolve(
-        `${__dirname}/../crates/zed/assets/themes/${theme.name}.json`
+        `${__dirname}/../../crates/zed/assets/themes/${theme.name}.json`
     );
     fs.writeFileSync(outPath, styleTreeJSON);
     console.log(`Generated ${outPath}`);
