+++ styles/src/styleTree/contactsPanel.ts
